requirements.txt
pandas
psycopg2-binary
tabulate
python-dotenvenv